Modified grid system to expand boxes in place
With the default Isotope layout, boxes in the last row open downwards past
the bottom of the container, and sometimes in other situations boxes move
significantly from their original position when expanded. I worked around
this by making the layout sorted, and calculating a sort order each time
a box is expanded.